add: add artist and title to top of saved lyrics
This is an important step for RAG search. If I prompt, for example,
"Give me a chorus for A Thousand Years by James Arthur", the lyrics
TXT file not containing the title and artist in the TXT does not give
this information to the LLM.

I tried using this prompt with an Ollama 3.1 8B, and it responded with
"Here's a chorus for A Thousand Years by Christina Perri (not James
Arthur". It's great that it gives out that disclaimer, and I also learnt
that to do a dumb RAG test, it's better to use smaller models which
have less information so that we know the retrieval part is actually
working or not.